🚧 export seekbar events
